+++ SLIDE/LSH.h
@@ -13,14 +13,14 @@ private:
   size_t _numBuckets;
   std::vector<size_t> _seeds;
 
-  void Add(int indices, int tableId, int id, bool unlimited);
+  void Add(size_t indices, size_t tableId, int id, bool unlimited);
 
 public:
   LSH(int K, int L, int RangePow);
   void clear();
-  void Add(const std::vector<int> &indices, int id, bool unlimited);
-  std::vector<int> hashesToIndex(const std::vector<int> &hashes) const;
+  void Add(const std::vector<size_t> &indices, int id, bool unlimited);
+  std::vector<size_t> hashesToIndex(const std::vector<int> &hashes) const;
   std::vector<const std::vector<int> *>
-  retrieveRaw(const std::vector<int> &indices) const;
+  retrieveRaw(const std::vector<size_t> &indices) const;
   ~LSH();
 };
